requirements/dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --no-index --output-file requirements/dev.txt requirements/base.txt requirements/testing.in requirements/dev.in
#
astroid==1.6.5            # via pylint
beautifulsoup4==4.7.1     # via webtest
bleach-whitelist==0.0.10
bleach==3.1.0
bumpversion==0.5.3
certifi==2019.3.9
chardet==3.0.4
click==7.0
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.3
cssselect==1.0.3          # via pyquery
django-appconf==1.0.3
django-axes==4.5.4
django-choices==1.6.2
django-cors-middleware==1.3.1
django-debug-toolbar==1.11
django-extensions==2.1.6
django-filter==2.1.0
django-ipware==2.1.0
django-markup==1.4.1
django-solo==1.1.3
django-webtest==1.9.4
django==2.1.8
djangorestframework-camel-case==1.0.3
djangorestframework==3.9.2
drf-nested-routers==0.91
drf-yasg==1.15.0
factory-boy==2.11.1
faker==1.0.4              # via factory-boy
freezegun==0.3.11
gemma-zds-client==0.9.4
idna==2.8
inflection==0.3.1
iso-639==0.4.5
isodate==0.6.0
isort==4.3.16
itypes==1.1.0
jinja2==2.10.1
lazy-object-proxy==1.3.1  # via astroid
lxml==4.3.3               # via pyquery
markdown==3.1
markupsafe==1.1.1
mccabe==0.6.1             # via pylint
mock==2.0.0
oyaml==0.9
pbr==5.1.3                # via mock
pep8==1.7.1
pip-tools==3.6.1
psycopg2==2.8
pyjwt==1.7.1
pylint==1.9.4
pyquery==1.4.0
python-dateutil==2.8.0    # via faker, freezegun
python-dotenv==0.10.1
pytz==2018.9
pyyaml==5.1
raven==6.10.0
requests==2.21.0
ruamel.yaml==0.15.90
six==1.12.0
soupsieve==1.9            # via beautifulsoup4
sqlparse==0.3.0           # via django-debug-toolbar
tblib==1.3.2
text-unidecode==1.2       # via faker
unidecode==1.0.23
uritemplate==3.0.0
urllib3==1.24.3
uwsgi==2.0.18
vng-api-common==0.46.2
waitress==1.2.1           # via webtest
webencodings==0.5.1
webob==1.8.5              # via webtest
webtest==2.0.33
wrapt==1.11.1             # via astroid
